RONDB-106: Batch execution improvements
This improvement makes it clear what the consistency of
batch executions are. The aim is to ensure that we can
always read our previous writes in the same transaction
even if we are using batched executions.

If there is a risk that we issue several operations in
the same row in overlapping batches we need to ensure that
all operations of this row is executed in the primary replica
first and using the LDM thread.

It isn't necessary for all operations in the batch to do this,
only those that have multiple operations on the same row.

We provide no special guarantees of what order executions on
different rows will have.

This means that for batch executions to be able to use
read backup and query threads we need to set a BatchSafe
flag on the TCKEYREQ signal. Otherwise only TCKEYREQ operations
with Exec flag will be considered for use by backup replicas and
query threads.