RONDB-106 Fix execute flag
The execute flag was never reset after being set in the first
interaction of the transaction. This had not so large issues
previously, only allowing TCKEYCONF to be sent a bit premature
in some cases.

However now being used to discover if we are executing a single
operation at a time, it is important to reset the execute flag
every time we get a TCKEYREQ from the NDB API and ensuring that
it is set again when the NDB API sends the last TCKEYREQ in a
batch.

Introduced a new flag, TF_SINGLE_EXEC_FLAG to indicate that the
interaction is using a single TCKEYREQ from the NDB API. This
single TCKEYREQ can still generate lots of other TCKEYREQ using
Reorg trigger, FK triggers, Fully replicated triggers, Unique
Key interactions. But those are all dependendant on the single
operation and is thus be ok to execute in any order. Also
the final Deferred trigger execution phase is ok to execute in
parallel, no interactions with NDB API happens at this point.