Bug#33872577 Main thread(tid:0) runs 'Out of job buffers' due to extra JB quota assigned.
Bug is a regression introduced by patch for:

   commit 367d9278a182676d8ba6a4d768ae9e3eae05c440
   Author: Mikael Ronström <mikael.ronstrom@oracle.com>
   Date:   Fri Jun 28 14:44:26 2019 +0200

    BUG#29887068

    Fixing some issues with main thread when restoring metadata where the
    heartbeats weren't sent properly due to signals in main thread
    executing for too long and thus no chance for heartbeats to be sent.

That patch, based on comments in the code and commit comments above, seemingly
intended to reduce maximum nunbers of Job buffer(JB) slots the  main-thread
signal execution could consume. Intention was to avoid the main-thread to
execute too long, as the comment above states.

However, instead of reducing the 'perjb' to minimum 5, which seems to have been
the intention, it was coded as 'MAX(perjb,5)'. Thus never letting perjb to drop
below '5', even if have reached the minimum (RESERVED) treshold of free JB's.

That allowed the main thread to continue execution when it should have yielded
the CPU and waited for more JBs to become available. That effectively broke the
'Out of job buffer' prevention mechanism for the main-thread.

Patch removes the offending piece of code, allowing a 'perjb=0' to be calculated.

We choose this solution over fixing the possible MIN/MAX mixup in breaking patch.
That patch consisted of multiple mechanism for preventing main-thread to run too long.
Apperently this brooken part of the patch has never worked as intended as it
didn't reduce the 'perjb' count. Moreover, always forcing 'perjb' count to 5 or
less for the main-thread, seems far too low as well. Normally it is
in the ~100 region, thus '5' would most certainly have introduced a
performance regression.

Change-Id: Ib723d93b66a60973e6a8917752f4fc300abf47bf