RONDB-119: Ensure that all nodes can be activated and hostname changed online
Previously a new command to activate, deactivate and changing hostnames
of nodes was introduced. This did change their status in the data nodes.

In this feature we add the capability to also change the status in API
and MGM server nodes. This is mainly interesting to ensure that API
nodes are informed of the changes to a data node. We do however also
provide the same information also for API and MGM nodes.

Since we don't know if all nodes are fully connected to the API nodes,
we send the information to all data nodes and each data node will send
to all connected API and MGM nodes. This means that an API node will
receive the same message multiple times. This should not be a concern
since the message execution is idempotent.