RONDB-120: Fixed a number of issues around IPv4
With configuration parameters on the data nodes and API nodes we will
always use the data node setting except for connections to MGM server.
Except when the MGM server connects as a client to the data node.
This makes the logic a bit complicated to figure out if we are using
IPv4 format or IPv6 for connect addresses.

Also fixed a bug in that no remote address was reported in transporters
table in ndbinfo. Fixed by ensuring that we get the connect address from
the first active node transporter rather than from the multi transporter.