RONDB-121: EnableRedoControl won't work when #LDMs > #LogParts and get_total_memory incorrect
Previously there was an assumption that the LDM thread had access
to log parts. Without this the LDM thread simply ignored any
attempt to set a proper disk write speed and instead used the
static checkpoint speed.

Now if #LDMs > #LogParts the LDM thread will look at all log parts
that it can use and add those together and divide it by the number
of LDM threads using those log parts. Thus the REDO log usage is
indicative of a combined REDO log usage of multiple LDM threads.

This required changing the calculation of percentage of REDO log used.
We calculate this in get_redo_stats instead where we rather check for
the REDO log which is the one with the highest redo percentage.

The function get_total_memory assumed that the memory usage among the
LDM threads was evenly distributed, this is not always correct. Replaced
by getting number of allocated pages in the LDM thread from DBTUP from
the variable m_pages_allocated which is already maintained. This led to
estimating the proposed disk write speed to low or too high and thus
the checkpoint time kept increasing with uneven use of the LDM threads.

Added an extra increase when the checkpoint time reaches beyond 1 minute,
2 minutes and 3 minutes to ensure that we try to retain the checkpoint
times fairly low. No more than 15% added in this manner.

When we execute in a high CPU state for an extended time we will use very
low disk write speeds. This would previously not be changed until we
reach 25% REDO fill level. Added now a check such that we slowly start
writing checkpoints at proposed disk write speeds if the lag is larger
than 1 GByte. Thus we allow for temporary high loads that create a lag
of 1 GByte, but after reaching a lag of 1 GByte we will try hard to keep
up with the proposed disk write speed even in the context of high load.

This is 1 GByte per LDM thread.