RONDB-754: Handled review comments and fixed RonDB config
The old default behaviour was that RonDB metadata cluster used
the same object as RonDB. This is now changed such that the
default config is that no metadata cluster exists. This means
that the metadata connection will simply be the first data
cluster connection. One gets this behaviour by setting
ConnectionPoolSize to 0 (default value for RonDB metadata
cluster connection). Setting it to 1 means that the cluster
connection need to be properly setup.

Enabled batchfeaturestore benchmark, it does however crash
for some reason (the Go benchmark code crashes that is).
Removed whitespaces from benchmark requests => 15% performance
improvement.