RONDB-754: Make allocation of request buffers more manageable
Instead of allocating a request buffer for every part of the
batch request I create buffers of twice the size, as long as
half of the buffer is still free I continue using this buffer
for the next request. This means that in almost all the cases
we will reuse the first buffer for all requests. Thus we only
use a lot of memory when it is actually needed.

Also fixed a lot of memory leak issues since almost every error
path failed to release both the request and the response buffers.
Probably the Go garbage collector handled this, but not in C++.