RONDB-754: Ensured that normally only one response buffer is required
Same idea as with Request buffer, reuse the current buffer until it is
filled to more than half.

Added a new configuration variable maxReqSize to set the maximum buffer
size used to parse the JSON request. This should not be the same as
the reqBufferSize that is how large request buffer we can store the
PK read request in, this variable is per PK request and not for the
entire batch size. Thus they are different.