RONDB-761: Make it possible to use dirtyWrite for INCR key
A dirty write still needs to update time tracking, thus we cannot
release the TC record before that is completed.

Dirty writes will commit on the primary replica and then commit
on backup replica. This means that it is possible to have the
data on the primary and backup to differ. However by using
writes all the time we are certain that if a difference happens,
it will immediately converge again after the next successful
write. One should avoid firing triggers when using dirty writes.

A new option OO_DIRTY_FLAG added for this purpose.