RONDB-671: Fix of STR_TO_INT64 instruction
This instruction uses strtoll to convert a string to a number.
It runs through the string until it finds a character that is
not a number.

It uses the heap memory that should have been updated by earlier
instructions. These instructions are known to contain a string
which is a number. However the string uses a length, so there
is no final character that is not the ASCII value of a number.

To avoid overwriting the heap memory we copy the string into
a local stack variable and set the byte after the string to
0. This ensures that the strtoll will never search beyond
the actual string.